finances-website/finances-api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
psycopg2-binary==2.9.10
pandas==2.2.3
python-dotenv==1.1.0
loguru==0.7.3
python-magic==0.4.27
